refactor(index.d.ts): remove ambient types
- if they are unused
- if they are used only once
- if they are duplicated
