+++ dotcom-rendering/src/types/matchReport.ts
@@ -1,0 +1,14 @@
+export type MatchReportType = {
+	id: string;
+	isResult: boolean;
+	homeTeam: TeamType;
+	awayTeam: TeamType;
+	competition: {
+		fullName: string;
+	};
+	isLive: boolean;
+	venue: string;
+	comments: string;
+	minByMinUrl: string;
+	reportUrl: string;
+};
